reworking code of the statement impl
