+++ jdbc-v2/src/main/java/com/clickhouse/jdbc/internal/SqlParser.java
@@ -10,9 +10,6 @@ import org.antlr.v4.runtime.tree.IterativeParseTreeWalker;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
 public class SqlParser {
 
     private static final Logger LOG = LoggerFactory.getLogger(SqlParser.class);
@@ -39,28 +36,6 @@ public class SqlParser {
         ParsedPreparedStatement parserListener = new ParsedPreparedStatement();
         IterativeParseTreeWalker.DEFAULT.walk(parserListener, parseTree);
         return parserListener;
-    }
-
-    private final static Pattern UNQUOTE_INDENTIFIER = Pattern.compile(
-            "^[\\\"`]?(.+?)[\\\"`]?$"
-    );
-
-    public static String unquoteIdentifier(String str) {
-        Matcher matcher = UNQUOTE_INDENTIFIER.matcher(str.trim());
-        if (matcher.find()) {
-            return matcher.group(1);
-        } else {
-            return str;
-        }
-    }
-
-    public static String escapeQuotes(String str) {
-        if (str == null || str.isEmpty()) {
-            return str;
-        }
-        return str
-                .replace("'", "\\'")
-                .replace("\"", "\\\"");
     }
 
     private static class ParserErrorListener extends BaseErrorListener {
